pyproject.toml
[tool.poetry]
name = "dog-breed-explorer"
version = "0.1.0"
description = "Pyne Dog Breed Explorer case"
authors = ["Frederik Møller Sørensen <frederik.moller2@hotmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^2.2.5"
requests = "^2.32.3"
google-auth = "^2.39.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
